refactor: dialog package 수정
